+++ posthog-server/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 1.1.0 - 2025-10-03
+
 - feat: `timestamp` can now be overridden when capturing an event ([#297](https://github.com/PostHog/posthog-android/issues/297))
 - feat: Add `groups`, `groupProperties`, `personProperties` overrides to feature flag methods ([#298](https://github.com/PostHog/posthog-android/issues/298))